% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_ACE.R
\name{decision_tree_ACE}
\alias{decision_tree_ACE}
\title{Decision Tree for ACE project}
\usage{
decision_tree_ACE(parameters, N.mc = 2, cost_dectree = "osNode_cost.Rds",
  health_dectree = "osNode_health.Rds", LTBI_pathString)
}
\arguments{
\item{cost_dectree}{Rds file names}

\item{health_dectree}{Rds file names}
}
\description{
Calculate decision tree expected costs and QALY loss
for N simulations
}
\examples{

}
