% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_level_ssize_sample.R
\name{patient_level_ssize_sample}
\alias{patient_level_ssize_sample}
\title{patient_level_ssize_sample}
\usage{
patient_level_ssize_sample(N = 1114, osNode, pathString, N.mc)
}
\arguments{
\item{osNode}{decision tree}

\item{n}{Initial patient cohort size}
}
\value{
Sampled patient subset size
}
\description{
patient_level_ssize_sample
}
