% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_LTBI_prob.R
\name{posterior_LTBI_prob}
\alias{posterior_LTBI_prob}
\title{Posterior LTBI probability}
\usage{
posterior_LTBI_prob(p_LTBI, t)
}
\arguments{
\item{p_LTBI}{Probability.}

\item{t}{Time (positive integer).}
}
\value{
Probability
}
\description{
Conditional on having 'survived' up to time \code{t}.
}
\examples{

(0.975*0.975*0.99*0.3)/(0.975*0.975*0.99*0.3 + 0.7)
posterior_LTBI_prob(0.3, 3)

}
