% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posttest_probs.R
\name{posttest_neg}
\alias{posttest_neg}
\title{posttest_neg}
\usage{
posttest_neg(prev, sens, spec)
}
\arguments{
\item{prev}{prevalence, probability}

\item{sens}{sensitivity, probability}

\item{spec}{specificity, probability}
}
\description{
P(Disease+|Test-)
}
