% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posttest_probs.R
\name{posttest_pos}
\alias{posttest_pos}
\title{posttest_pos}
\usage{
posttest_pos(prev, sens, spec)
}
\arguments{
\item{prev}{prevalence, probability}

\item{sens}{sensitivity, probability}

\item{spec}{specificity, probability}
}
\description{
P(Disease+|Test+)
}
