% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posttest_probs.R
\name{posttest_pos_aggregate}
\alias{posttest_pos_aggregate}
\title{posttest_pos_aggregate}
\usage{
posttest_pos_aggregate(n_pos, n_neg, sens, spec, prev)
}
\arguments{
\item{n_pos}{Number of positives}

\item{n_neg}{Number of negatives}

\item{sens}{Sensitivity}

\item{spec}{Specificity}

\item{prev}{Prevalence}
}
\description{
##TODO: check maths
}
\details{
See Modeling in Medical Decision Making: A Bayesian Approach
Giovanni Parmigiani book example

Currently, only computes with pmf for prevalence of two probabilities
i.e. p(x=1)=prevalence
so then the likelihood reeduces to sensitivity and specificity
if using a distn for prevalence eg beta/uniform
then will need to integrate denominator.
}
\examples{

posttest_pos_aggregate(n_pos = 10, n_neg = 10, sens = 0.9, spec = 0.9, prev = 0.1)

}
