% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_pathway_prob_sample.R
\name{total_pathway_prob_sample}
\alias{total_pathway_prob_sample}
\title{Total pathway probability sample}
\usage{
total_pathway_prob_sample(osNode,
  pathString = "LTBI screening cost/LTBI/Agree to Screen/Sensitivity",
  N.mc = 2)
}
\arguments{
\item{osNode}{Decision tree object.}

\item{pathString}{Route from root to leaf. Node names separated by /.}

\item{N.mc}{Number of samples.}
}
\value{
Vector of probabilities.
}
\description{
Total pathway probability sample
}
\examples{

}
